// tools/conformance/src/modules/nestjs-dataloader-module-generator/nestjs-dataloader-module-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsDataloaderModuleGeneratorArguments {
  options: NestjsDataloaderModuleGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsDataloaderModuleGeneratorOptions {
  name: string;
  project?: string;
}